refactor: combine 2 util files into one
